+++ engine/lite/embeddedDeploymentManager/src/main/scala/pl/touk/nussknacker/streaming/embedded/EmbeddedDeploymentManager.scala
@@ -6,13 +6,14 @@ import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
 import pl.touk.nussknacker.engine.api._
 import pl.touk.nussknacker.engine.api.deployment._
-import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
+import pl.touk.nussknacker.engine.api.deployment.simple.{SimpleProcessStateDefinitionManager, SimpleStateStatus}
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.api.queryablestate.QueryableClient
 import pl.touk.nussknacker.engine.lite.api.runtimecontext.LiteEngineRuntimeContextPreparer
-import pl.touk.nussknacker.engine.lite.kafka.KafkaTransactionalScenarioInterpreter
+import pl.touk.nussknacker.engine.lite.kafka.{KafkaTransactionalScenarioInterpreter, TaskStatus}
 import pl.touk.nussknacker.engine.lite.metrics.dropwizard.{DropwizardMetricsProviderFactory, LiteEngineMetrics}
 import pl.touk.nussknacker.engine.graph.EspProcess
+import pl.touk.nussknacker.engine.lite.kafka.TaskStatus.TaskStatus
 import pl.touk.nussknacker.engine.marshall.ScenarioParser
 import pl.touk.nussknacker.engine.{DeploymentManagerProvider, ModelData, TypeSpecificInitialData}
 import sttp.client.{NothingT, SttpBackend}
@@ -61,6 +62,8 @@ class EmbeddedDeploymentManager(modelData: ModelData, engineConfig: Config,
   private val metricRegistry = LiteEngineMetrics.prepareRegistry(engineConfig)
 
   private val contextPreparer = new LiteEngineRuntimeContextPreparer(new DropwizardMetricsProviderFactory(metricRegistry))
+
+  override def processStateDefinitionManager: ProcessStateDefinitionManager = EmbeddedProcessStateDefinitionManager
 
   @volatile private var interpreters: Map[ProcessName, ScenarioInterpretationData] = {
     val deployedScenarios = Await.result(processingTypeDeploymentService.getDeployedScenarios, retrieveDeployedScenariosTimeout)
@@ -111,10 +114,19 @@ class EmbeddedDeploymentManager(modelData: ModelData, engineConfig: Config,
 
   override def findJobStatus(name: ProcessName): Future[Option[ProcessState]] = Future.successful {
     interpreters.get(name).map { interpreterData =>
-      ProcessState(interpreterData.deploymentId, SimpleStateStatus.Running,
-        Some(interpreterData.processVersion), processStateDefinitionManager)
+      ProcessState(
+        deploymentId = interpreterData.deploymentId,
+        status = statusMapping.getOrElse(interpreterData.scenarioInterpreter.status(), SimpleStateStatus.NotDeployed),
+        version = Some(interpreterData.processVersion),
+        definitionManager = processStateDefinitionManager
+      )
     }
   }
+
+  private val statusMapping: Map[TaskStatus, StateStatus] = Map(
+    TaskStatus.Running -> SimpleStateStatus.Running,
+    TaskStatus.Restarting -> EmbeddedStateStatus.Restarting
+  )
 
   override def close(): Unit = {
     interpreters.values.foreach(_.scenarioInterpreter.close())