+++ engine/lite/embeddedDeploymentManager/src/main/scala/pl/touk/nussknacker/streaming/embedded/EmbeddedStateStatus.scala
@@ -1,0 +1,8 @@
+package pl.touk.nussknacker.streaming.embedded
+
+import pl.touk.nussknacker.engine.api.deployment._
+import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
+
+object EmbeddedStateStatus  {
+  val Restarting: StateStatus = NotEstablishedStateStatus("Restarting")
+}