+++ engine/lite/embeddedDeploymentManager/src/test/scala/pl/touk/nussknacker/streaming/embedded/EmbeddedDeploymentManagerTest.scala
@@ -1,14 +1,14 @@
 package pl.touk.nussknacker.streaming.embedded
 
 import com.typesafe.config.ConfigFactory
-import com.typesafe.config.ConfigValueFactory.fromAnyRef
+import com.typesafe.config.ConfigValueFactory.{fromAnyRef, fromMap}
 import io.circe.Json.{fromString, obj}
 import org.scalatest.{FunSuite, Matchers}
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.ProcessVersion
 import pl.touk.nussknacker.engine.api.deployment.TestProcess.{ExpressionInvocationResult, TestData}
-import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.deployment._
+import pl.touk.nussknacker.engine.api.deployment.simple.SimpleStateStatus
 import pl.touk.nussknacker.engine.api.process.ProcessName
 import pl.touk.nussknacker.engine.api.runtimecontext.IncContextIdGenerator
 import pl.touk.nussknacker.engine.build.EspProcessBuilder
@@ -23,13 +23,13 @@ import pl.touk.nussknacker.engine.spel.Implicits._
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.util.ThreadUtils
 import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
-import pl.touk.nussknacker.test.{FailingContextClassloader, PatientScalaFutures}
+import pl.touk.nussknacker.test.{FailingContextClassloader, VeryPatientScalaFutures}
 
 import java.util.UUID
 import scala.concurrent.ExecutionContext.Implicits._
 import scala.jdk.CollectionConverters.mapAsJavaMapConverter
 
-class EmbeddedDeploymentManagerTest extends FunSuite with KafkaSpec with Matchers with PatientScalaFutures {
+class EmbeddedDeploymentManagerTest extends FunSuite with KafkaSpec with Matchers with VeryPatientScalaFutures {
 
   case class FixtureParam(deploymentManager: DeploymentManager, modelData: ModelData, inputTopic: String, outputTopic: String) {
     def deployScenario(scenario: EspProcess): Unit = {
@@ -51,6 +51,12 @@ class EmbeddedDeploymentManagerTest extends FunSuite with KafkaSpec with Matcher
       .withValue("auto.offset.reset", fromAnyRef("earliest"))
       .withValue("exceptionHandlingConfig.topic", fromAnyRef("errors"))
       .withValue("components.kafka.enabled", fromAnyRef(true))
+      .withValue("kafka.kafkaProperties", fromMap(Map[String, Any](
+//        This timeout controls how long the kafka producer initialization in pl.touk.nussknacker.engine.lite.kafka.KafkaSingleScenarioTaskRun.init.
+//        So it has to be set to a reasonably low value for the restarting test to finish before ScalaFutures patience runs out.
+        "max.block.ms" -> 1000,
+        "default.api.timeout.ms" -> 1000
+      ).asJava))
 
     val modelData = LocalModelData(configToUse, new EmptyProcessConfigCreator)
     val deploymentService = new ProcessingTypeDeploymentServiceStub(initiallyDeployedScenarios)
@@ -110,6 +116,36 @@ class EmbeddedDeploymentManagerTest extends FunSuite with KafkaSpec with Matcher
     manager.findJobStatus(name).futureValue shouldBe None
   }
 
+  test("Set status to restarting when scenario fails and back to running when the problems are fixed") {
+    val fixture@FixtureParam(manager, _, inputTopic, outputTopic) = prepareFixture()
+
+    val name = ProcessName("testName")
+    val scenario = EspProcessBuilder
+      .id(name.value)
+      .source("source", "kafka-json", "topic" -> s"'$inputTopic'")
+      .emptySink("sink", "kafka-json", "topic" -> s"'$outputTopic'", "value" -> "#input")
+
+    wrapInFailingLoader {
+      fixture.deployScenario(scenario)
+    }
+
+    manager.findJobStatus(name).futureValue.map(_.status) shouldBe Some(SimpleStateStatus.Running)
+
+    kafkaZookeeperServer.kafkaServer.shutdown()
+    kafkaZookeeperServer.kafkaServer.awaitShutdown()
+
+    eventually {
+      val jobStatus = manager.findJobStatus(name).futureValue
+      jobStatus.map(_.status) shouldBe Some(EmbeddedStateStatus.Restarting)
+      jobStatus.map(_.allowedActions).get should contain only (ProcessActionType.Cancel)
+    }
+
+    kafkaZookeeperServer.kafkaServer.startup()
+
+    eventually {
+      manager.findJobStatus(name).futureValue.map(_.status) shouldBe Some(SimpleStateStatus.Running)
+    }
+  }
 
   test("Redeploys scenario") {
     val fixture@FixtureParam(manager, _, inputTopic, outputTopic) = prepareFixture()
@@ -164,7 +200,7 @@ class EmbeddedDeploymentManagerTest extends FunSuite with KafkaSpec with Matcher
 
     val testData = TestData(new ModelDataTestInfoProvider(modelData).generateTestData(scenario.metaData,
         scenario.roots.head.data.asInstanceOf[Source], 2).get, 2)
-    
+
     val results = wrapInFailingLoader {
       manager.test(name, ProcessMarshaller.toJson(ProcessCanonizer.canonize(scenario)).noSpaces, testData, identity[Any]).futureValue
     }